Remove redundant code and simplify
Previously, two event listeners were created in global. It was not clear
what the event listeners were doing. Now a function has been created to
add those event listeners, which makes it more clear what they are
doing. Additionally, the code does not need to be duplicated for the
second search form.